fixbug volume sửa lại UI

See merge request cloudsoft/oneportal-frontend!280

+++ apps/app-smart-cloud/src/app/pages/volume/component/action/delete-volume/delete-volume.component.ts
@@ -12,14 +12,17 @@ export class DeleteVolumeComponent {
   @Input() region: number
   @Input() project: number
   @Input() volumeId: number
+  @Input() volumeName: string
   @Output() onOk = new EventEmitter()
   @Output() onCancel = new EventEmitter()
 
   isLoading: boolean = false
   isVisible: boolean = false
 
-  constructor(@Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService,
-              private notification: NzNotificationService,
+  value: string
+  isInput: boolean = false
+
+  constructor(private notification: NzNotificationService,
               private volumeService: VolumeService) {
   }
 
@@ -30,28 +33,43 @@ export class DeleteVolumeComponent {
   handleCancel() {
     this.isVisible = false
     this.isLoading = false
+    this.isInput = false
+    this.value = null
     this.onCancel.emit()
+  }
+
+  onInputChange(value) {
+    this.value = value;
   }
 
   handleOk() {
     this.isLoading = true
-    this.volumeService.deleteVolume(this.volumeId).subscribe(data => {
-      if(data) {
-        this.isLoading = false
-        this.isVisible = false
-        this.notification.success('Thành công', 'Xóa Volume thành công')
-        this.onOk.emit(data)
-      } else {
-        console.log('data', data)
+    if (this.value == this.volumeName) {
+      this.isInput = false
+      this.volumeService.deleteVolume(this.volumeId).subscribe(data => {
+        if (data) {
+          this.isLoading = false
+          this.isVisible = false
+          this.notification.success('Thành công', 'Xóa Volume thành công')
+          this.onOk.emit(data)
+        } else {
+          console.log('data', data)
+          this.isLoading = false
+          this.isVisible = false
+          this.notification.error('Thất bại', 'Xóa Volume thất bại')
+        }
+      }, error => {
+        console.log('error', error)
         this.isLoading = false
         this.isVisible = false
         this.notification.error('Thất bại', 'Xóa Volume thất bại')
-      }
-    }, error => {
-      console.log('error', error)
+      })
+    } else {
+      this.isInput = true
       this.isLoading = false
-      this.isVisible = false
-      this.notification.error('Thất bại', 'Xóa Volume thất bại')
-    })
+    }
   }
+
 }
+
+
